starting the browserfiy proccess
