+++ dokomoforms/static/widgets.js
@@ -1,640 +1,12 @@
-var NEXT = 1;
-var PREV = -1;
 var ON = true;
 var OFF = false;
 var NUM_FAC = 256;
 var FAC_RAD = 2; //in KM
 
-var App = {
-    unsynced: [], // unsynced surveys
-    facilities: [], // revisit facilities
-    unsynced_facilities: {}, // new facilities
-    location: {},
-    submitter_name: ''
-};
+var getNearbyFacilities = require('./facilities.js').getNearbyFacilities;
+var objectID = require('./facilities.js').objectID();
 
-App.init = function(survey) {
-    var self = this;
-    self.survey = new Survey(survey.survey_id, 
-            survey.survey_version, 
-            survey.questions, 
-            survey.survey_metadata, 
-            survey.survey_title, 
-            survey.created_on,
-            survey.last_updated);
-
-    var start_loc = survey.survey_metadata.location 
-        || {'lat': 40.8138912, 'lon': -73.9624327}; // defaults to nyc
-
-    self.facilities = JSON.parse(localStorage.facilities || "[]");
-    self.submitter_name = localStorage.name || "";
-    self.submitter_email = localStorage.email || "";
-    
-    // Init if unsynced is undefined
-    if (!localStorage.unsynced) {
-        localStorage.unsynced = JSON.stringify({});
-    }
-    // Load up any unsynced submissions
-    App.unsynced = JSON.parse(localStorage.unsynced)[self.survey.id] || []; 
-
-    // Load up any unsynced facilities
-    App.unsynced_facilities = 
-        JSON.parse(localStorage.unsynced_facilities || "{}");
-
-    // Load any facilities
-    App.facilities = JSON.parse(localStorage.facilities || "{}");
-    if (JSON.stringify(App.facilities) === "{}" && navigator.onLine) {
-        // See if you can get some new facilities
-        getNearbyFacilities(start_loc.lat, start_loc.lon, 
-            FAC_RAD, // Radius in km 
-            NUM_FAC, // limit
-            null// what to do with facilities 
-        );
-    }
-
-    // AppCache updates
-    window.applicationCache.addEventListener('updateready', function() {
-        alert('app updated, reloading...');
-        window.location.reload();
-    });
-
-    $('header')
-        .on('click', '.menu', function(e) {
-            $('header').toggleClass('title-extended');
-            $('.title_menu').toggle();
-        });
-
-    $('header')
-        .on('click', '.menu_clear', function(e) {
-            $('header').toggleClass('title-extended');
-            $('.title_menu').toggle();
-
-            localStorage.clear();
-            App.splash();
-            App.message('All survey data erased.', 'Surveys Nuked', 'message-warning');
-        });
-
-    $('header')
-        .on('click', '.menu_restart', function(e) {
-            $('header').toggleClass('title-extended');
-            $('.title_menu').toggle();
-
-            App.survey.clearState();
-            App.splash();
-            App.message('Active survey has been cleared.', 'Survey Reset', 'message-warning');
-        });
-
-    $('header')
-        .on('click', '.menu_save', function(e) {
-            $('header').toggleClass('title-extended');
-            $('.title_menu').toggle();
-
-            App.survey.saveState();
-            App.message('Active survey has been saved.', 'Survey Saved', 'message-success');
-        });
-        
-    App.splash();
-};
-
-App.sync = function() {
-    var self = this;
-    //$('.submit_modal')[0].click(); // Pop Up Submitting Modal
-    self.countdown = App.unsynced.length; //JS is single threaded no race condition counter
-    self.count = App.unsynced.length; //JS is single threaded no race condition counter
-    var endSync = function() {
-        //$('.submit_modal')[0].click(); // Remove submitting modal;
-        App.splash();
-        if (!App.unsynced.length) {
-            App.message('All ' + self.count + ' surveys synced succesfully.', 'Survey Synced', 'message-success');
-        } else {
-            App.message(App.unsynced.length + ' survey(s) failed to sync succesfully. Please try again later.', 'Survey Sync Failed', 'message-error');
-        }
-    };
-    _.each(App.unsynced, function(survey) {
-        App.submit(survey, 
-            function(survey) { 
-                console.log('done');
-                // Has to be there
-                var idx = App.unsynced.indexOf(survey);
-                App.unsynced.splice(idx, 1);
-                var unsynced = JSON.parse(localStorage.unsynced); 
-                unsynced[self.survey.id] = App.unsynced;
-                localStorage['unsynced'] = JSON.stringify(unsynced);
-                --self.countdown; 
-                
-                if (self.countdown === 0) {  
-                    endSync();
-                }
-            },
-
-            function(survey) { 
-                console.log('fail');
-                --self.countdown; 
-
-                if (self.countdown === 0) {
-                    endSync();
-                }
-            } 
-        );
-    });
-
-    // Facilities junk
-    _.map(App.unsynced_facilities, function(facility) {
-        postNewFacility(facility); 
-    });
-};
-
-App.message = function(text, title, style) {
-    // Shows a message to user
-    // E.g. "Your survey has been submitted"
-    $('.message_btn')[0].click();
-    
-    $('.modal_header').empty()
-        //XXX Look into doing this in a more clean way
-        .removeClass('message-primary')
-        .removeClass('message-error')
-        .removeClass('message-warning')
-        .removeClass('message-success')
-        .addClass(style)
-        .text(title);
-
-
-    // Message text region
-    $('.message')
-        .text(text);
-};
-
-App.splash = function() {
-    $('header').removeClass('title-extended');
-    $('.title_menu').hide();
-    var self = this;
-    var survey = self.survey;
-    $('.overlay').hide(); // Always remove overlay after moving
-
-    // Update page nav bar
-    var barnav  = $('.bar-nav');
-    var barnavHTML = $('#template_nav__splash').html();
-    var barnavTemplate = _.template(barnavHTML);
-    var compiledHTML = barnavTemplate({
-        'org': survey.org,
-    });
-    
-    barnav.empty()
-        .html(compiledHTML);
-
-    var barfoot = $('.bar-footer');
-    barfoot.removeClass('bar-footer-extended');
-    barfoot.removeClass('bar-footer-super-extended');
-    barfoot.css("height", "");
-
-    var barfootHTML = $('#template_footer__splash').html();
-    var barfootTemplate = _.template(barfootHTML);
-    compiledHTML = barfootTemplate({
-    });
-
-    barfoot.empty()
-        .html(compiledHTML)
-        .find('.start_btn')
-        .one('click', function() {
-            // Render first question
-            App.survey.render(App.survey.first_question);
-        });
-
-
-    // Set up content
-    var content = $('.content');
-    content.removeClass('content-shrunk');
-    content.removeClass('content-super-shrunk');
-
-    var splashHTML = $('#template_splash').html();
-    var splashTemplate = _.template(splashHTML);
-    compiledHTML = splashTemplate({
-        'survey': survey,
-        'online': navigator.onLine,
-        'name': App.submitter_name,
-        'unsynced': App.unsynced,
-        'unsynced_facilities': App.unsynced_facilities
-    });
-
-    content.empty()
-        .data('index', 0)
-        .html(compiledHTML)
-        .find('.sync_btn')
-        .one('click', function() {
-            if (navigator.onLine) {
-                App.sync();
-                // Reload page to update template values
-                App.splash();
-            } else {
-                App.message('Please connect to the internet first.', 'Connection Error', 'message-warning');
-            }
-        });
-};
-
-App.submit = function(survey, done, fail) {
-    function getCookie(name) {
-        var r = document.cookie.match("\\b" + name + "=([^;]*)\\b");
-        return r ? r[1] : undefined;
-    }
-    
-    // Update submit time
-    survey.submission_time = new Date().toISOString();
-
-    $.ajax({
-        url: '/api/surveys/'+survey.survey_id+'/submit',
-        type: 'POST',
-        contentType: 'application/json',
-        processData: false,
-        data: JSON.stringify(survey),
-        headers: {
-            "X-XSRFToken": getCookie("_xsrf")
-        },
-        dataType: 'json',
-        success: function() {
-            done(survey);
-        },
-        error: function() {
-            fail(survey);
-        }
-    });
-
-    console.log('synced submission:', survey);
-    console.log('survey', '/api/surveys/'+survey.survey_id+'/submit');
-}
-
-function Survey(id, version, questions, metadata, title, created_on, last_updated) {
-    var self = this;
-    this.id = id;
-    this.questions = questions;
-    this.metadata = metadata;
-    this.author = metadata.author || 'anon';
-    this.org = metadata.organization || 'independant';
-    this.version = version;
-    this.title = title;
-    this.created_on = new Date(created_on).toDateString();
-    this.last_updated = new Date(last_updated).toDateString();
-
-    // Load answers from localStorage
-    var answers = JSON.parse(localStorage[this.id] || '{}');
-    //console/g.log(answers);
-    _.each(self.questions, function(question) {
-        question.answer = answers[question.question_id] || [];
-        // Set next pointers
-        question.next = self.getQuestion(question.question_to_sequence_number);
-    });
-
-    App.location = answers['location'] || {};
-
-    // Know where to start, and number
-    self.current_question = self.questions[0];
-    self.lowest_sequence_number = self.current_question.sequence_number;
-    self.first_question = self.current_question;
-
-    // Now that you know order, you can set prev pointers
-    var curr_q = self.current_question;
-    var prev_q = null;
-    do {
-        curr_q.prev = prev_q;
-        prev_q = curr_q;
-        curr_q = curr_q.next;
-    } while (curr_q);
-    
-}
-
-// Search by sequence number instead of array pos
-Survey.prototype.getQuestion = function(seq) {
-    var self = this;
-    for(var i = 0; i < self.questions.length; i++) {
-        if (self.questions[i].sequence_number === seq) {
-            return self.questions[i];
-        }
-    }
-
-    return null;
-};
-
-// Answer array may have elements even if answer[0] is undefined
-// Return a non empty response or an empty one if none found
-Survey.prototype.getFirstResponse = function(question) {
-    for (var i = 0; i < question.answer.length; i++) {
-        var answer = question.answer[i];
-        if (answer && typeof answer.response !== 'undefined') {
-            return answer
-        }
-    }
-
-    return {'response': null, 'is_type_exception': false, 'metadata': null};
-};
-
-// Choose next question, deals with branching and back/forth movement
-Survey.prototype.next = function(offset) {
-    var self = this;
-
-    var next_question = offset === PREV ? this.current_question.prev : this.current_question.next;
-    var index = $('.content').data('index');
-
-    var first_answer = this.getFirstResponse(this.current_question); 
-    var first_response = first_answer.response;
-    var first_is_type_exception = first_answer.is_type_exception;
-    var first_metadata = first_answer.metadata;
-
-    // Backward at first question
-    if (index === self.lowest_sequence_number && offset === PREV) {
-        App.splash();
-        return;
-    }
-
-    // Backwards at submit page
-    if (index === this.questions.length + 1 && offset === PREV) {
-        // Going backwards at submit means render ME;
-        next_question = this.current_question;
-    } 
-    
-    // Normal forward
-    if (offset === NEXT) {
-        // Is it a valid response?
-        bad_answers = [];
-        this.current_question.answer.forEach(function(resp) {
-            if (resp && resp.failed_validation)
-                bad_answers.push(resp);
-        });
-
-        if (bad_answers.length) {
-            App.message(bad_answers.length 
-            + ' response(s) found not valid for question type: ' 
-            + self.current_question.type_constraint_name, 'Survey Response Error', 'message-error');
-            return;
-        }
-
-        // Are you required?
-        if (this.current_question.logic.required && (first_response === null)) {
-            App.message('Survey requires this question to be completed.','Survey Required Response',  'message-error');
-            return;
-        }
-
-        // Is the only response and empty is other response?
-        if (first_is_type_exception && !first_response) {
-            App.message('Please provide a reason before moving on.', 'Survey Missing Reason', 'message-error');
-            return;
-        }
-
-        // Check if question was a branching question
-        if (this.current_question.branches && (first_response !== null)) {
-            var branches = this.current_question.branches;
-            for (var i=0; i < branches.length; i++) {
-                if (branches[i].question_choice_id === first_response) {
-                    next_question = self.getQuestion(branches[i].to_sequence_number);
-                    // update pointers
-                    self.current_question.next = next_question;
-                    next_question.prev = self.current_question; 
-                    break; // only one set of ptrs ever needed updating
-                }
-            }
-        }
-    }
-
-    self.saveState();
-    self.render(next_question);
-};
-
-// Render template for given question
-Survey.prototype.render = function(question) {
-    $('header').removeClass('title-extended');
-    $('.title_menu').hide();
-
-    var self = this;
-    $('.overlay').hide(); // Always remove overlay after moving
-
-    // Clear any interval events
-    if (Widgets.interval) {
-        window.clearInterval(Widgets.interval);
-        Widgets.interval = null;
-    }
-
-    var index = question ? question.sequence_number : this.questions.length + 1;
-
-    // Update navs
-    var barnav  = $('.bar-nav');
-    var barnavHTML = $('#template_nav').html();
-    var barnavTemplate = _.template(barnavHTML);
-    var compiledHTML = barnavTemplate({
-        'index': index,
-        'total': this.questions.length + 1,
-    });
-
-    barnav.empty()
-        .html(compiledHTML);
-
-    // Update footer
-    var barfoot = $('.bar-footer');
-    var barfootHTML;
-    var barfootTemplate;
-
-    barfoot.removeClass('bar-footer-extended');
-    barfoot.removeClass('bar-footer-super-extended');
-    barfoot.css("height", "");
-
-    // Update content
-    var content = $('.content');
-    var widgetHTML;
-    var widgetTemplate;
-    
-    if (question) {
-
-        // Add the next button
-        barfootHTML = $('#template_footer').html();
-        barfootTemplate = _.template(barfootHTML);
-        compiledHTML = barfootTemplate({
-            'other_text': question.logic.other_text
-        });
-
-        barfoot.empty()
-            .html(compiledHTML);
-
-        // Show widget
-        widgetHTML = $('#widget_' + question.type_constraint_name).html();
-        widgetTemplate = _.template(widgetHTML);
-        compiledHTML = widgetTemplate({question: question});
-        self.current_question = question;
-
-        // Render question
-        content.removeClass('content-shrunk');
-        content.removeClass('content-super-shrunk');
-        content.empty()
-            .data('index', index)
-            .html(compiledHTML)
-            .scrollTop(); //XXX: Ignored in chrome ...
-        
-        // Attach widget events
-        Widgets[question.type_constraint_name](question, content, barfoot);
-
-    } else {
-        // Add submit button
-        barfootHTML = $('#template_footer__submit').html();
-        barfootTemplate = _.template(barfootHTML);
-        compiledHTML = barfootTemplate({
-        });
-
-        barfoot.empty()
-            .html(compiledHTML)
-            .find('.submit_btn')
-                .one('click', function() {
-                    self.submit();
-                });
-
-        // Show submit page
-        widgetHTML = $('#template_submit').html();
-        widgetTemplate = _.template(widgetHTML);
-        compiledHTML = widgetTemplate({
-                'name': App.submitter_name,
-                'email': App.submitter_email
-        });
-
-        // Render submit page
-        content.removeClass('content-shrunk');
-        content.removeClass('content-super-shrunk');
-        content.empty()
-            .data('index', index)
-            .html(compiledHTML)
-            .find('.name_input')
-            .keyup(function() {
-                App.submitter_name = this.value;
-                localStorage.name = App.submitter_name;
-            });
-
-        content
-            .find('.email_input')
-            .keyup(function() {
-                App.submitter_email = this.value;
-                localStorage.email = App.submitter_email;
-            });
-    }
-    
-    // Update nav
-    $('.page_nav__progress')
-        .text((index) + ' / ' + (this.questions.length + 1));
-    
-    // Page navigation
-    $('.page_nav__prev, .page_nav__next').click(function() {
-        var offset = $(this).hasClass('page_nav__prev') ? PREV : NEXT;
-        self.next(offset);
-    });
-    
-
-};
-
-Survey.prototype.saveState = function() {
-    var self = this;
-    var answers = {};
-
-    // Save answers locally 
-    _.each(self.questions, function(question) {
-        answers[question.question_id] = question.answer;
-    });
-    answers['location'] = App.location;
-
-    // Save answers in storage
-    localStorage[self.id] = JSON.stringify(answers);
-}
-
-Survey.prototype.clearState = function() {
-    var self = this;
-
-    // Clear answers locally 
-    _.each(self.questions, function(question) {
-        question.answer = [];
-    });
-    App.location = {};
-
-    // Clear answers in storage
-    localStorage[self.id] = JSON.stringify({});
-}
-
-Survey.prototype.submit = function() {
-    var self = this;
-
-    // Prepare POST request
-    var survey_answers = [];
-    self.questions.forEach(function(q) {
-        //console.log('q', q);
-        q.answer.forEach(function(ans, ind) {
-
-            if (ans == null) {
-                return;
-            }
-
-            var response =  ans.response;
-            var is_type_exception = ans.is_type_exception || false;
-            var metadata = ans.metadata || {};
-            var is_new_facility = metadata.is_new; //XXX: Should I remove this is new marking?
-
-            if (response == null) { 
-                return;
-            }
-
-            if (is_new_facility) {
-                // Record this new facility for Revisit s)ubmission
-                App.unsynced_facilities[response.id] = {
-                    'name': metadata.name, 'uuid': response.id, 
-                    'properties' : {'sector': metadata.sector},
-                    'coordinates' : [response.lon, response.lat]
-                };
-
-                // Store it in facilities as well
-                App.facilities[response.id] = App.unsynced_facilities[response.id];
-            }
-
-            survey_answers.push({
-                question_id: q.question_id,
-                answer: response,
-                answer_metadata: metadata,
-                is_type_exception: is_type_exception
-            });
-
-        });
-    });
-
-    var data = {
-        submitter: App.submitter_name || "anon",
-        submitter_email: App.submitter_email || "anon@anon.org",
-        survey_id: self.id,
-        answers: survey_answers,
-        save_time: new Date().toISOString()
-    };
-
-    console.log('saved submission:', data);
-    
-    // Don't post with no replies
-    if (JSON.stringify(survey_answers) === '[]') {
-      // Not doing instantly to make it seem like App tried reaaall hard
-      setTimeout(function() {
-            App.message('Saving failed, No questions answered in Survey!', 'Survey Empty Submission', 'message-warning');
-            App.splash();
-      }, 1000);
-      return;
-    } 
-
-    // Save Revisit data 
-    localStorage.setItem("facilities", 
-            JSON.stringify(App.facilities));
-
-    localStorage.setItem("unsynced_facilities", 
-            JSON.stringify(App.unsynced_facilities));
-    
-    // Clear State
-    self.clearState();
-
-    // Save Submission data
-    App.unsynced.push(data);
-    var unsynced = JSON.parse(localStorage.unsynced); 
-    unsynced[self.id] = App.unsynced;
-    localStorage['unsynced'] = JSON.stringify(unsynced);
-    //App.message('Please remember to sync submissions when connected to the internet.', 'Survey Saved', 'message-primary');
-    App.splash();
-
-
-};
-
+//TODO:Remove refernce to App
 
 var Widgets = {
     interval: null,
@@ -1408,86 +780,4 @@ Widgets.facility = function(question, page, footer) {
 
 };
 
-
-/* -------------------------- Revisit Stuff Below ----------------------------*/
-function getNearbyFacilities(lat, lng, rad, lim, cb) {
-    var url = "http://staging.revisit.global/api/v0/facilities.json"; 
-
-    // Revisit ajax req
-    //console/g.log("MADE EXTERNAL REVISIT QUERY");
-    $.get(url,{
-            near: lat + "," + lng,
-            rad: rad,
-            limit: lim,
-            //sortDesc: "updatedAt",
-            fields: "name,uuid,coordinates,properties:sector", //filters results to include just those three fields,
-        },
-        function(data) {
-            facilities = {};
-            data.facilities.forEach(function(facility) {
-                facilities[facility.uuid] = facility;
-            });
-            // Add in our unsynced ones as well
-            Object.keys(App.unsynced_facilities).forEach(function(uuid) {
-                facilities[uuid] = App.unsynced_facilities[uuid];
-            });
-
-            App.facilities = facilities;
-            localStorage.setItem("facilities", JSON.stringify(facilities));
-            if (cb) {
-                cb(facilities); //drawFacillities callback probs
-            }
-        }
-    );
-}
-
-function postNewFacility(facility) {
-    var url = "http://staging.revisit.global/api/v0/facilities.json"; 
-
-    $.ajax({
-        url: url,
-        type: 'POST',
-        contentType: 'application/json',
-        data: JSON.stringify(facility),
-        processData: false,
-        dataType: 'json',
-        success: function() {
-            //App.message('Facility Added!', 'message-primary');
-            // If posted, we don't an unsynced reference to it anymore
-            delete App.unsynced_facilities[facility.uuid];
-        },
-        
-        headers: {
-            "Authorization": "Basic " + btoa("dokomoforms" + ":" + "password")
-                //DONT DO THIS XXX XXX
-        },
-
-        error: function() {
-            //App.message('Facility submission failed, will try again later.', 'message-warning');
-        },
-        
-        complete: function() {
-            localStorage.setItem("facilities", 
-                    JSON.stringify(App.facilities));
-
-            localStorage.setItem("unsynced_facilities", 
-                    JSON.stringify(App.unsynced_facilities));
-
-        }
-    });
-}
-
-// Def not legit but hey
-function objectID() {
-    return 'xxxxxxxxxxxxxxxxxxxxxxxx'.replace(/[x]/g, function() {
-        var r = Math.random()*16|0;
-        return r.toString(16);
-    });
-}
-
-var exports = exports || {} //XXX: Just to silence console; 
-exports.App = App;
-exports.Survey = Survey;
 exports.Widgets = Widgets; 
-exports.getNearbyFacilities = getNearbyFacilities;
-exports.postNewFacility = postNewFacility;
